+++ slashCommands/global/about.js
@@ -63,7 +63,7 @@ class about {
 		}
 		buttonRow.addComponents(
 			new ButtonBuilder().setStyle(ButtonStyle.Link).setLabel("Status Updates").setURL("https://discord.gg/3mRbVCjvmz"),
-			new ButtonBuilder().setStyle(ButtonStyle.Link).setLabel("Website").setURL("https://www.febkosq8.me/Jappan"),
+			new ButtonBuilder().setStyle(ButtonStyle.Link).setLabel("Website").setURL("https://jappan.febkosq8.me"),
 		);
 		const uptime = await ClientHandler.getClientUptime();
 		const owner = await ClientHandler.getClientUser("407819516513484801");