+++ ChangeLog.md
@@ -1,9 +1,27 @@
+## v1.4.3
+
+- discord.js, @discord-player, etc version bump
+- Prefix commands only gets triggered after a pre check
+- Added GetChannel admin command to fetch a channel using guild id and channel id
+- Audit logger now converts the object to string the event while saving to db
+- GuildList admin command now sends the embed(s) in pairs of 20 servers and also includes the timestamp of join and the owner user id
+- Remove all reference to user discriminators
+- Doppelganger now takes your globalName instead of nickname
+- AuditHandler now includes try catch blocks to avoid random uncaught exceptions
+- AuditHandler now skips events that was triggered by the bot itself
+- Play command now truncates the autocomplete item name(s) if more than 100 characters
+- Added listener for unhandledRejection
+- MongoDB disconnection event now waits 10 seconds to propagate. This handles the case where the MongoDB instance was switching to a new cluster
+
+#
+
 ## v1.4.2
 
 - discord.js, @discord-player, etc version bump
+- discord.js, @discord-player, etc version bump
 - Switched to useMainPlayer from UseMasterPlayer for discord-player
 - Switched to mediaplex from opusscript
-- We no longer send the send on interaction fail event
+- We no longer send the error stack interaction fail event
 - AdminGuildList now also checks if bot has Admin role
 - Removed ffmpeg from npm and changed it to docker
 - Fixed AuditHandler throwing error when missing some data